requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
together-python==0.1.7
